+++ worker/src/types.d.ts
@@ -50,6 +50,7 @@ type Bindings = {
     USER_ROLES: string | UserRole[] | undefined
     DOMAIN_LABELS: string | string[] | undefined
     PASSWORDS: string | string[] | undefined
+    ADMIN_USERNAMES: string | string[] | undefined
     ADMIN_PASSWORDS: string | string[] | undefined
     DISABLE_ADMIN_PASSWORD_CHECK: string | boolean | undefined
     JWT_SECRET: string